Set up a rough architecture
The interface Game represents an app which can be launched by the
engine. LabExplorer implements that interface.
In the long run, the game will provide the engine with information about
the scene, which the engine will then use to calculate physics, render
the scene on the screen, and maybe handle extra stuff like user input,
sound or even, dare I say, netcode.